collector/requirements.txt
httpx>=0.27,<1.0
feedparser>=6.0,<7.0
beautifulsoup4>=4.12,<5.0
pyyaml>=6.0,<7.0
redis>=5.0,<6.0
schedule>=1.2,<2.0
